benchmarks/requirements.txt
# no versions since benchmarks should always run with the most recent version of each package

python-dateutil
trafaret
django
djangorestframework
marshmallow
valideer
attr
cattrs
cerberus
